+++ app/Http/Controllers/LoginController.php
@@ -2,19 +2,28 @@
 
 namespace App\Http\Controllers;
 
+use App\Login;
 use App\Notifications\SecureLoginCode;
 use App\SecureLogin;
+use App\User;
+use Illuminate\Auth\Events\Failed;
+use Illuminate\Auth\SessionGuard;
+use Illuminate\Foundation\Auth\ThrottlesLogins;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\ValidationException;
+use Lab404\AuthChecker\Services\AuthChecker;
 
 class LoginController extends Controller
 {
+    use ThrottlesLogins;
+
+    public $maxAttempts = 5;
+    public $decayMinutes = 1;
+
     public function __construct()
     {
         $this->middleware('guest')->except('logout');
-
-        $this->middleware('throttle:5,1')->only('check');
     }
 
     public function form()
@@ -29,20 +38,56 @@ class LoginController extends Controller
             'password' => 'required|string',
         ]);
 
+        if ($this->hasTooManyLoginAttempts($request)) {
+            $this->fireLockoutEvent($request);
+            $this->sendLockoutResponse($request);
+        }
+
+        if (true === app('shaarli')->getSecureLogin()) {
+            return $this->checkWithSecureLogin($request, $validated);
+        }
+
+        return $this->checkWithoutSecureLogin($request, $validated);
+    }
+
+    protected function checkWithoutSecureLogin(Request $request, array $validated)
+    {
         if (Auth::guard()->attempt($validated, $request->filled('remember'))) {
             $request->session()->regenerate();
 
-            if (true === app('shaarli')->getSecureLogin()) {
-                $user = Auth::guard()->user();
-                $secure = SecureLogin::createForUser($user);
-                $user->notify(new SecureLoginCode($secure));
-                Auth::logout();
-
-                return redirect()->route('login.secure', $secure);
-            }
-
             return redirect()->intended('/');
         }
+
+        $this->incrementLoginAttempts($request);
+
+        throw ValidationException::withMessages([
+            'email' => [__("Invalid credentials")],
+        ]);
+    }
+
+    protected function checkWithSecureLogin(Request $request, array $validated)
+    {
+        /** @var SessionGuard $guard */
+        $guard = Auth::guard();
+
+        if ($guard->validate($validated)) {
+            /** @var User $user */
+            $user = $guard->getLastAttempted();
+
+            /** @var AuthChecker $checker */
+            $checker = app('authchecker');
+            $device = $checker->findOrCreateUserDeviceByAgent($user);
+            $checker->createUserLoginForDevice($user, $device, Login::TYPE_2FA);
+
+            $secure = SecureLogin::createForUser($user);
+            $user->notify(new SecureLoginCode($secure));
+
+            return redirect()->route('login.secure', $secure);
+        }
+
+        event(new Failed('web', $guard->getLastAttempted(), $validated));
+
+        $this->incrementLoginAttempts($request);
 
         throw ValidationException::withMessages([
             'email' => [__("Invalid credentials")],
@@ -56,5 +101,10 @@ class LoginController extends Controller
         $request->session()->invalidate();
 
         return redirect('/');
+    }
+
+    public function username(): string
+    {
+        return 'email';
     }
 }