+++ app/Notifications/SecureLoginCode.php
@@ -29,8 +29,8 @@ class SecureLoginCode extends Notification
     {
         return (new MailMessage)
                 ->subject(__("Secure your login"))
-                ->line(__("Please use the following code to access your account."))
-                ->action($this->secure->code, sprintf('%s?code=%s', route('login.secure', $this->secure), $this->secure->code));
+                ->line(__("Please use the following code :code to access your account.", ['code' => $this->secure->code]))
+                ->action(__("Confirm login"), sprintf('%s?code=%s', route('login.secure', $this->secure), $this->secure->code));
     }
 
     public function toArray($notifiable)